--- a/composer.json
+++ b/composer.json
@@ -14,8 +14,7 @@
     "sort-packages": true
   },
   "require": {
-    "php": "^7.0",
-    "localheinz/token": "0.2.2"
+    "php": "^7.0"
   },
   "require-dev": {
     "codeclimate/php-test-reporter": "0.4.4",