This PR

* [x] inlines `localheinz/token`

### Before

```
+-------------+------------------------------+--------+--------+------+-----+------------+---------------+---------------+---------------+---------------+---------+--------+----------+
| benchmark   | subject                      | groups | params | revs | its | mem_peak   | best          | mean          | mode          | worst         | stdev   | rstdev | diff     |
+-------------+------------------------------+--------+--------+------+-----+------------+---------------+---------------+---------------+---------------+---------+--------+----------+
| ClassyBench | benchBaseline                |        | []     | 10   | 1   | 5,487,504b | 28,247.100μs  | 28,247.100μs  | 28,247.100μs  | 28,247.100μs  | 0.000μs | 0.00%  | 0.00%    |
| ClassyBench | benchConstructsFromDirectory |        | []     | 10   | 1   | 5,984,800b | 137,113.600μs | 137,113.600μs | 137,113.600μs | 137,113.600μs | 0.000μs | 0.00%  | +385.41% |
+-------------+------------------------------+--------+--------+------+-----+------------+---------------+---------------+---------------+---------------+---------+--------+----------+
```

### After

```
+-------------+------------------------------+--------+--------+------+-----+------------+--------------+--------------+--------------+--------------+---------+--------+---------+
| benchmark   | subject                      | groups | params | revs | its | mem_peak   | best         | mean         | mode         | worst        | stdev   | rstdev | diff    |
+-------------+------------------------------+--------+--------+------+-----+------------+--------------+--------------+--------------+--------------+---------+--------+---------+
| ClassyBench | benchBaseline                |        | []     | 10   | 1   | 5,486,904b | 28,509.300μs | 28,509.300μs | 28,509.300μs | 28,509.300μs | 0.000μs | 0.00%  | 0.00%   |
| ClassyBench | benchConstructsFromDirectory |        | []     | 10   | 1   | 5,765,112b | 45,173.100μs | 45,173.100μs | 45,173.100μs | 45,173.100μs | 0.000μs | 0.00%  | +58.45% |
+-------------+------------------------------+--------+--------+------+-----+------------+--------------+--------------+--------------+--------------+---------+--------+---------+
```